:sparkles: v3: new Start method for app
