+++ client_test.go
@@ -28,13 +28,17 @@ func Test_Client_Invalid_URL(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.SendString(c.Hostname())
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	a := Get("http://example.com\r\n\r\nGET /\r\n\r\n")
 
@@ -65,13 +69,17 @@ func Test_Client_Get(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.SendString(c.Hostname())
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	for i := 0; i < 5; i++ {
 		a := Get("http://example.com")
@@ -91,13 +99,17 @@ func Test_Client_Head(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.SendString(c.Hostname())
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	for i := 0; i < 5; i++ {
 		a := Head("http://example.com")
@@ -117,14 +129,18 @@ func Test_Client_Post(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Post("/", func(c *Ctx) error {
 		return c.Status(StatusCreated).
 			SendString(c.FormValue("foo"))
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	for i := 0; i < 5; i++ {
 		args := AcquireArgs()
@@ -151,13 +167,17 @@ func Test_Client_Put(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Put("/", func(c *Ctx) error {
 		return c.SendString(c.FormValue("foo"))
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	for i := 0; i < 5; i++ {
 		args := AcquireArgs()
@@ -184,13 +204,17 @@ func Test_Client_Patch(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Patch("/", func(c *Ctx) error {
 		return c.SendString(c.FormValue("foo"))
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	for i := 0; i < 5; i++ {
 		args := AcquireArgs()
@@ -217,14 +241,18 @@ func Test_Client_Delete(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Delete("/", func(c *Ctx) error {
 		return c.Status(StatusNoContent).
 			SendString("deleted")
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	for i := 0; i < 5; i++ {
 		args := AcquireArgs()
@@ -248,13 +276,17 @@ func Test_Client_UserAgent(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.Send(c.Request().Header.UserAgent())
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	t.Run("default", func(t *testing.T) {
 		for i := 0; i < 5; i++ {
@@ -390,13 +422,17 @@ func Test_Client_Agent_Host(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.SendString(c.Hostname())
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	a := Get("http://1.1.1.1:8080").
 		Host("example.com").
@@ -486,13 +522,17 @@ func Test_Client_Agent_Custom_Response(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.SendString("custom")
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	for i := 0; i < 5; i++ {
 		a := AcquireAgent()
@@ -523,13 +563,17 @@ func Test_Client_Agent_Dest(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.SendString("dest")
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	t.Run("small dest", func(t *testing.T) {
 		dest := []byte("de")
@@ -591,9 +635,13 @@ func Test_Client_Agent_RetryIf(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	a := Post("http://example.com").
 		RetryIf(func(req *Request) bool {
@@ -698,7 +746,7 @@ func Test_Client_Agent_MultipartForm(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Post("/", func(c *Ctx) error {
 		utils.AssertEqual(t, "multipart/form-data; boundary=myBoundary", c.Get(HeaderContentType))
@@ -710,7 +758,11 @@ func Test_Client_Agent_MultipartForm(t *testing.T) {
 		return c.Send(c.Request().Body())
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	args := AcquireArgs()
 
@@ -753,7 +805,7 @@ func Test_Client_Agent_MultipartForm_SendFiles(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Post("/", func(c *Ctx) error {
 		utils.AssertEqual(t, "multipart/form-data; boundary=myBoundary", c.Get(HeaderContentType))
@@ -780,7 +832,11 @@ func Test_Client_Agent_MultipartForm_SendFiles(t *testing.T) {
 		return c.SendString("multipart form files")
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	for i := 0; i < 5; i++ {
 		ff := AcquireFormFile()
@@ -884,14 +940,18 @@ func Test_Client_Agent_Timeout(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		time.Sleep(time.Millisecond * 200)
 		return c.SendString("timeout")
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	a := Get("http://example.com").
 		Timeout(time.Millisecond * 50)
@@ -910,13 +970,17 @@ func Test_Client_Agent_Reuse(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.SendString("reuse")
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	a := Get("http://example.com").
 		Reuse()
@@ -951,13 +1015,17 @@ func Test_Client_Agent_InsecureSkipVerify(t *testing.T) {
 
 	ln = tls.NewListener(ln, serverTLSConf)
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.SendString("ignore tls")
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	code, body, errs := Get("https://" + ln.Addr().String()).
 		InsecureSkipVerify().
@@ -980,13 +1048,17 @@ func Test_Client_Agent_TLS(t *testing.T) {
 
 	ln = tls.NewListener(ln, serverTLSConf)
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.SendString("tls")
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	code, body, errs := Get("https://" + ln.Addr().String()).
 		TLSConfig(clientTLSConf).
@@ -1002,7 +1074,7 @@ func Test_Client_Agent_MaxRedirectsCount(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		if c.Request().URI().QueryArgs().Has("foo") {
@@ -1014,7 +1086,11 @@ func Test_Client_Agent_MaxRedirectsCount(t *testing.T) {
 		return c.SendString("redirect")
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	t.Run("success", func(t *testing.T) {
 		a := Get("http://example.com?foo").
@@ -1048,7 +1124,7 @@ func Test_Client_Agent_Struct(t *testing.T) {
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", func(c *Ctx) error {
 		return c.JSON(data{true})
@@ -1058,7 +1134,11 @@ func Test_Client_Agent_Struct(t *testing.T) {
 		return c.SendString(`{"success"`)
 	})
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	t.Run("success", func(t *testing.T) {
 		t.Parallel()
@@ -1127,11 +1207,15 @@ func testAgent(t *testing.T, handler Handler, wrapAgent func(agent *Agent), exce
 
 	ln := fasthttputil.NewInmemoryListener()
 
-	app := New(Config{DisableStartupMessage: true})
+	app := New()
 
 	app.Get("/", handler)
 
-	go func() { utils.AssertEqual(t, nil, app.Listener(ln)) }()
+	go func() {
+		utils.AssertEqual(t, nil, app.Start(ln, StartConfig{
+			DisableStartupMessage: true,
+		}))
+	}()
 
 	c := 1
 	if len(count) > 0 {
